backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.24.2
marshmallow-sqlalchemy==1.1.1
mysql-connector-python==9.1.0
packaging==24.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
flask_login==0.6.3
flask_dance[mysql]
oauthlib
requests-oauthlib
python-dotenv
pytest-flask
gunicorn 
typing_extensions
flask_migrate